Fixed error and success screens not appearing after provider authorization
