+++ frontend/src/routeTree.gen.ts
@@ -22,6 +22,7 @@ import { Route as AuthenticatedDashboardRouteImport } from './routes/_authentica
 import { Route as AuthenticatedUsersIndexRouteImport } from './routes/_authenticated/users/index'
 import { Route as UsersUserIdPairRouteImport } from './routes/users/$userId/pair'
 import { Route as AuthenticatedUsersUserIdRouteImport } from './routes/_authenticated/users/$userId'
+import { Route as UsersUserIdPairIndexRouteImport } from './routes/users/$userId/pair.index'
 import { Route as UsersUserIdPairSuccessRouteImport } from './routes/users/$userId/pair.success'
 import { Route as UsersUserIdPairErrorRouteImport } from './routes/users/$userId/pair.error'
 
@@ -90,6 +91,11 @@ const AuthenticatedUsersUserIdRoute =
     path: '/$userId',
     getParentRoute: () => AuthenticatedUsersRoute,
   } as any)
+const UsersUserIdPairIndexRoute = UsersUserIdPairIndexRouteImport.update({
+  id: '/',
+  path: '/',
+  getParentRoute: () => UsersUserIdPairRoute,
+} as any)
 const UsersUserIdPairSuccessRoute = UsersUserIdPairSuccessRouteImport.update({
   id: '/success',
   path: '/success',
@@ -116,6 +122,7 @@ export interface FileRoutesByFullPath {
   '/users/': typeof AuthenticatedUsersIndexRoute
   '/users/$userId/pair/error': typeof UsersUserIdPairErrorRoute
   '/users/$userId/pair/success': typeof UsersUserIdPairSuccessRoute
+  '/users/$userId/pair/': typeof UsersUserIdPairIndexRoute
 }
 export interface FileRoutesByTo {
   '/': typeof IndexRoute
@@ -127,10 +134,10 @@ export interface FileRoutesByTo {
   '/settings': typeof AuthenticatedSettingsRoute
   '/widget/connect': typeof WidgetConnectRoute
   '/users/$userId': typeof AuthenticatedUsersUserIdRoute
-  '/users/$userId/pair': typeof UsersUserIdPairRouteWithChildren
   '/users': typeof AuthenticatedUsersIndexRoute
   '/users/$userId/pair/error': typeof UsersUserIdPairErrorRoute
   '/users/$userId/pair/success': typeof UsersUserIdPairSuccessRoute
+  '/users/$userId/pair': typeof UsersUserIdPairIndexRoute
 }
 export interface FileRoutesById {
   __root__: typeof rootRouteImport
@@ -149,6 +156,7 @@ export interface FileRoutesById {
   '/_authenticated/users/': typeof AuthenticatedUsersIndexRoute
   '/users/$userId/pair/error': typeof UsersUserIdPairErrorRoute
   '/users/$userId/pair/success': typeof UsersUserIdPairSuccessRoute
+  '/users/$userId/pair/': typeof UsersUserIdPairIndexRoute
 }
 export interface FileRouteTypes {
   fileRoutesByFullPath: FileRoutesByFullPath
@@ -167,6 +175,7 @@ export interface FileRouteTypes {
     | '/users/'
     | '/users/$userId/pair/error'
     | '/users/$userId/pair/success'
+    | '/users/$userId/pair/'
   fileRoutesByTo: FileRoutesByTo
   to:
     | '/'
@@ -178,10 +187,10 @@ export interface FileRouteTypes {
     | '/settings'
     | '/widget/connect'
     | '/users/$userId'
-    | '/users/$userId/pair'
     | '/users'
     | '/users/$userId/pair/error'
     | '/users/$userId/pair/success'
+    | '/users/$userId/pair'
   id:
     | '__root__'
     | '/'
@@ -199,6 +208,7 @@ export interface FileRouteTypes {
     | '/_authenticated/users/'
     | '/users/$userId/pair/error'
     | '/users/$userId/pair/success'
+    | '/users/$userId/pair/'
   fileRoutesById: FileRoutesById
 }
 export interface RootRouteChildren {
@@ -305,6 +315,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AuthenticatedUsersUserIdRouteImport
       parentRoute: typeof AuthenticatedUsersRoute
     }
+    '/users/$userId/pair/': {
+      id: '/users/$userId/pair/'
+      path: '/'
+      fullPath: '/users/$userId/pair/'
+      preLoaderRoute: typeof UsersUserIdPairIndexRouteImport
+      parentRoute: typeof UsersUserIdPairRoute
+    }
     '/users/$userId/pair/success': {
       id: '/users/$userId/pair/success'
       path: '/success'
@@ -354,11 +371,13 @@ const AuthenticatedRouteWithChildren = AuthenticatedRoute._addFileChildren(
 interface UsersUserIdPairRouteChildren {
   UsersUserIdPairErrorRoute: typeof UsersUserIdPairErrorRoute
   UsersUserIdPairSuccessRoute: typeof UsersUserIdPairSuccessRoute
+  UsersUserIdPairIndexRoute: typeof UsersUserIdPairIndexRoute
 }
 
 const UsersUserIdPairRouteChildren: UsersUserIdPairRouteChildren = {
   UsersUserIdPairErrorRoute: UsersUserIdPairErrorRoute,
   UsersUserIdPairSuccessRoute: UsersUserIdPairSuccessRoute,
+  UsersUserIdPairIndexRoute: UsersUserIdPairIndexRoute,
 }
 
 const UsersUserIdPairRouteWithChildren = UsersUserIdPairRoute._addFileChildren(
